Change main app directories structure
