+++ src/app/App.tsx
@@ -5,7 +5,7 @@ import { RouteComponentProps } from "react-router";
 
 import { Footer, MainMenu, MetaConsumer, OverlayManager } from "../components";
 import { isPath } from "../core/utils";
-import { orderConfirmationUrl, Routes } from "../routes";
+import { orderConfirmationUrl, Routes } from "./routes";
 
 const App: React.FC<RouteComponentProps> = ({
   history: {
