+++ src/checkout/index.ts
@@ -1,0 +1,1 @@
+export { default as CheckoutApp } from "./CheckoutApp";
