--- src/routes/index.ts
@@ -1,2 +1,0 @@
-export * from "./Routes";
-export { default as Routes } from "./Routes";
