+++ src/userAccount/views/OrderDetails/Page.tsx
@@ -11,7 +11,7 @@ import {
   OrderByToken_orderByToken_lines,
 } from "./types/OrderByToken";
 
-import { orderHistoryUrl } from "../../../routes";
+import { orderHistoryUrl } from "../../../app/routes";
 
 const extractOrderLines = (
   lines: Array<OrderById_order_lines | OrderByToken_orderByToken_lines>
