+++ src/views/OrderConfirmation/View.tsx
@@ -3,9 +3,9 @@ import "./scss/index.scss";
 import React from "react";
 import { generatePath, Link, RouteComponentProps } from "react-router-dom";
 
+import { guestOrderDetailsUrl } from "../../app/routes";
 import { Button, NotFound } from "../../components";
 import { BASE_URL } from "../../core/config";
-import { guestOrderDetailsUrl } from "../../routes";
 
 const View: React.FC<RouteComponentProps> = ({
   history: {
